+++ src/CP/Utilities/CoreUtilities.php
@@ -71,6 +71,7 @@ class CoreUtilities
             ->routes(function ($router) {
                 $router->get('refresh', [LicensingController::class, 'refresh'])->name('refresh');
                 $router->post('mint', [LicensingController::class, 'mint'])->name('mint');
+                $router->post('fresh', [LicensingController::class, 'fresh'])->name('fresh');
             });
 
         if (config('statamic.git.enabled') && Statamic::pro()) {