+++ src/Providers/ConsoleServiceProvider.php
@@ -60,7 +60,6 @@ class ConsoleServiceProvider extends ServiceProvider
         Commands\ImportUsers::class,
         Commands\ProEnable::class,
         Commands\SiteFreshKey::class,
-        Commands\SiteRotateKey::class,
     ];
 
     public function boot()