+++ tests/Licensing/SiteLicenseTest.php
@@ -24,11 +24,22 @@ class SiteLicenseTest extends TestCase
     }
 
     #[Test]
-    public function it_prefers_the_site_key_over_the_legacy_license_key()
+    public function it_prefers_the_legacy_license_key_over_the_site_key()
     {
         config([
             'statamic.system.site_key' => 'site_abcdefghijklmnopqrstuvwxyz',
             'statamic.system.license_key' => 'aRadLicenseKey42',
+        ]);
+
+        $this->assertEquals('aRadLicenseKey42', $this->license()->key());
+    }
+
+    #[Test]
+    public function it_uses_the_site_key_when_there_is_no_legacy_license_key()
+    {
+        config([
+            'statamic.system.site_key' => 'site_abcdefghijklmnopqrstuvwxyz',
+            'statamic.system.license_key' => null,
         ]);
 
         $this->assertEquals('site_abcdefghijklmnopqrstuvwxyz', $this->license()->key());